Code for Intro to VR workshop escape room minigame
This commit contains the code for the Intro to VR workshop escape room mini-game. This contains the basic process flow of the game which starts from the initial TTS prompt followed by using a ball to drop a chandelier. The droppped chandelier reveals a book. The revealed book is then placed in a hidden book slot in the bookcase. Placing the book in the slot causes the bookcase to open. Following this, a TTS is played signifying the end of the escape room game. The ending sequence is yet to be added. Also need to add the instructions for building this.